Update mcd_file.py
The before and after ablation images are PNGs, and usually one will not want to do any processing on this data.

The current functionality retrieves this data as a numpy array, due to which the properties attached to the png file are lost (e.g. timestamp when it was created).

The code has been modified to maintain the normal behavior, but if the "raw" boolean flag is specified during the request of these files, then the raw png data is returned (as opposed to the numpy array).

One can easily save this data to a png file, and retain all the metadata associated with the png file.